requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
gunicorn>=21.0.0
